winkernel: move CanUseWinKernelProxier to util
CanUseWinKernelProxier is a kube-proxy thing and kpng doesn't use so far.
However, we might need to use, keep it for now in util.
